REFACTOR steps/1-prijmy + fixed warnings + TEST fix
