+++ src/pages/odklad/osobne-udaje.tsx
@@ -1,4 +1,4 @@
-import { validate, OsobneUdaje } from '../../plugins/odklad/2-osobne-udaje/osobne-udaje.tsx'
+import { validate, OsobneUdaje } from '../../plugins/odklad/2-osobne-udaje/osobne-udaje'
 
 export { validate }
 export default OsobneUdaje
