+++ src/pages/odklad/pokracovat.tsx
@@ -1,3 +1,3 @@
-import { ContinuePage } from '../../plugins/odklad/4-pokracovat/pokracovat.tsx'
+import { ContinuePage } from '../../plugins/odklad/4-pokracovat/pokracovat'
 
 export default ContinuePage
