+++ src/pages/odklad/prijmy-zo-zahranicia.tsx
@@ -1,4 +1,4 @@
-import { validate, PrijmyZoZahranicia } from '../../plugins/odklad/1-prijmy/prijmy-zo-zahranicia.tsx'
+import { validate, PrijmyZoZahranicia } from '../../plugins/odklad/1-prijmy/prijmy-zo-zahranicia'
 
 export { validate }
 export default PrijmyZoZahranicia
